webApi/db+mqtt/app/requirements.txt
psycopg2-binary
sqlalchemy
python-dotenv
fastapi[standard]
bcrypt
